refactor(test): 테스트 분리
